+++ CHANGELOG.md
@@ -13,4 +13,5 @@
  - Breaking: Rename `Invoke-SalesforceApexFile` to `Invoke-SalesforceApex` for consistency. Update scripts and imports accordingly.
  - Breaking: Rename `Test-Salesforce` to `Test-SalesforceApex` to clarify scope. Update scripts and imports accordingly.
  - Breaking: Rename psfdx-logs functions with plural suffix from `*SalesforceLogs` to `*SalesforceDebugLogs` (`Watch-`, `Get-`, `Export-`). Update scripts and imports accordingly.
- - Breaking: Rename `Get-SalesforceLog` to `Get-SalesforceDebugLog`.
+- Breaking: Rename `Get-SalesforceLog` to `Get-SalesforceDebugLog`.
+ - Breaking: Rename `Convert-SalesforceLog` to `Convert-SalesforceDebugLog`.